Hash1: Make comment more precisely describe how to emulate logic from shell
The shell command previously given as an example emits paths starting
with `./`, and without an added prefix. Also, previously, `sha256` was
run against the sorted list of filenames, not against the _contents_ of
those files (`xargs` is needed to perform the latter).

This does introduce a GNUism; working with POSIX `find` might involve
adding a `sed` pipeline element.