go.mod: ignore cyclic dependency for tagging
tools and mod have a cyclic dependency, which needs to be broken during
the upgrading/tagging process. This direction is due to the use of txtar
in the tests for the zip package, which is much less important than the
other direction.

For golang/go#48523.

Change-Id: Id5006280b0bc6b39f191b0281d542d251d80fff5
Reviewed-on: https://go-review.googlesource.com/c/mod/+/439335
Reviewed-by: Bryan Mills <bcmills@google.com>
Run-TryBot: Heschi Kreinick <heschi@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>