modfile: improve directory path detection and error text consistency
An error text suggests a directory path needs to start with ./ or ../
if it's a relative path, but in reality relative paths with .\ and ..\
prefix (such as those that are used on Windows) are also accepted.

Furthermore, a relative path like ./ or ../ is fine, as are ./. and
../., but the cleaner and shorter equivalent relative paths . and ..
are reported as if they're not directory paths (even though a module
path cannot consist of nothing but dots).

Fix those inconsistencies and make IsDirectoryPath report true on "."
and ".." paths as expected, and make its documentation clear that
a path like "sub/dir", despite being a relative path, is interpreted
as a module path.

For golang/go#60572.

Change-Id: I8fa4a2c66bc83a1ccafc453b96f3bb33dc222cd1
Reviewed-on: https://go-review.googlesource.com/c/mod/+/500335
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Bryan Mills <bcmills@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>